+++ paddle/phi/api/lib/kernel_dispatch.cc
@@ -56,7 +56,7 @@ BackendSet GetTensorBackendSet(const phi::TensorBase& t) {
   if (HasAllocation(t) && t.place().GetType() != AllocationType::UNDEFINED) {
     BackendSet backend_set(phi::TransToPhiBackend(t.place()));
     switch (t.layout()) {
-      case DataLayout::MKLDNN:
+      case DataLayout::ONEDNN:
         backend_set = backend_set | BackendSet(Backend::ONEDNN);
         break;
       default:
