+++ paddle/phi/kernels/onednn/log_softmax_kernel.cc
@@ -14,8 +14,8 @@
 
 #include "paddle/phi/kernels/log_softmax_kernel.h"
 
-#include "paddle/fluid/platform/mkldnn_reuse.h"
 #include "paddle/phi/backends/onednn/onednn_context.h"
+#include "paddle/phi/backends/onednn/onednn_reuse.h"
 #include "paddle/phi/common/bfloat16.h"
 #include "paddle/phi/common/place.h"
 #include "paddle/phi/core/kernel_registry.h"
@@ -23,16 +23,15 @@
 namespace phi {
 
 template <typename T>
-class LogSoftmaxMKLDNNHandler
-    : public paddle::platform::
-          MKLDNNHandlerNoCachingT<T, dnnl::logsoftmax_forward> {
+class LogSoftmaxOneDNNHandler
+    : public funcs::OneDNNHandlerNoCachingT<T, dnnl::logsoftmax_forward> {
  public:
-  LogSoftmaxMKLDNNHandler(const dnnl::engine mkldnn_engine,
+  LogSoftmaxOneDNNHandler(const dnnl::engine onednn_engine,
                           Place cpu_place,
                           const DenseTensor& x,
                           const int axis)
-      : paddle::platform::MKLDNNHandlerNoCachingT<T, dnnl::logsoftmax_forward>(
-            mkldnn_engine, cpu_place) {
+      : funcs::OneDNNHandlerNoCachingT<T, dnnl::logsoftmax_forward>(
+            onednn_engine, cpu_place) {
     this->AcquireForwardPrimitiveDescriptor(
         dnnl::prop_kind::forward_inference, x.mem_desc(), axis);
   }
@@ -43,11 +42,11 @@ void LogSoftmaxKernel(const Context& dev_ctx,
                       const DenseTensor& x,
                       int axis,
                       DenseTensor* out) {
-  const auto& mkldnn_engine = dev_ctx.GetEngine();
+  const auto& onednn_engine = dev_ctx.GetEngine();
   axis = axis >= 0 ? axis : x.dims().size() + axis;
 
-  LogSoftmaxMKLDNNHandler<T> handler(
-      mkldnn_engine, dev_ctx.GetPlace(), x, axis);
+  LogSoftmaxOneDNNHandler<T> handler(
+      onednn_engine, dev_ctx.GetPlace(), x, axis);
 
   auto src_memory_p = handler.AcquireSrcMemory(&x);
   auto dst_memory_p = handler.AcquireDstMemory(out);
